Remove Utils tests and refactor EJBCA certificate code
